Initial commit
Ported the addon to Vanilla 1.12.1.  It compiles and runes but there is
an issue with the networking code.  Seems like Ace2 (and every addon
that uses it), does not like the way we setup our networked code.  Might
be worth looking into converting to ace2 for the sake of compliance with
other popular addons.